Using unsigned int for numChildren parameters in build API.

+++ include/embree2/rtcore_builder.h
@@ -73,13 +73,13 @@ struct RTCORE_ALIGN(32) RTCBuildPrimitive
 RTCORE_API RTCBVH rtcNewBVH(RTCDevice device);
 
 /*! Callback to create a node. */
-typedef void* (*RTCCreateNodeFunc) (RTCThreadLocalAllocator allocator, size_t numChildren, void* userPtr);
+typedef void* (*RTCCreateNodeFunc) (RTCThreadLocalAllocator allocator, unsigned int numChildren, void* userPtr);
 
 /*! Callback to set the pointer to all children. */
-typedef void  (*RTCSetNodeChildrenFunc) (void* nodePtr, void** children, size_t numChildren, void* userPtr);
+typedef void  (*RTCSetNodeChildrenFunc) (void* nodePtr, void** children, unsigned int numChildren, void* userPtr);
 
 /*! Callback to set the bounds of all children. */
-typedef void  (*RTCSetNodeBoundsFunc) (void* nodePtr, const struct RTCBounds** bounds, size_t numChildren, void* userPtr);
+typedef void  (*RTCSetNodeBoundsFunc) (void* nodePtr, const struct RTCBounds** bounds, unsigned int numChildren, void* userPtr);
 
 /*! Callback to create a leaf node. */
 typedef void* (*RTCCreateLeafFunc) (RTCThreadLocalAllocator allocator, const struct RTCBuildPrimitive* prims, size_t numPrims, void* userPtr);
